LPS-176997 - module moved
