+++ modules/dxp/apps/analytics/analytics-audit-header-test/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Analytics Audit Header Test
+Bundle-SymbolicName: com.liferay.analytics.audit.header.test
+Bundle-Version: 1.0.0
