+++ modules/dxp/apps/analytics/analytics-audit-header/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Analytics Audit Header
+Bundle-SymbolicName: com.liferay.analytics.audit.header
+Bundle-Version: 1.0.0
