+++ modules/dxp/apps/analytics/analytics-audit-header/build.gradle
