--- modules/dxp/apps/osb/osb-analytics/osb-analytics-audit-header-test/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay OSB Analytics Audit Header Test
-Bundle-SymbolicName: com.liferay.osb.analytics.audit.header.test
-Bundle-Version: 1.0.0
