--- modules/dxp/apps/osb/osb-analytics/osb-analytics-audit-header-test/build.gradle
@@ -1,5 +1,0 @@
-dependencies {
-	testIntegrationCompile project(":apps:portal-configuration:portal-configuration-test-util")
-	testIntegrationCompile project(":dxp:apps:osb:osb-analytics:osb-analytics-audit-header")
-	testIntegrationCompile project(":test:arquillian-extension-junit-bridge")
-}
