--- modules/dxp/apps/osb/osb-analytics/osb-analytics-audit-header/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay OSB Analytics Audit Header
-Bundle-SymbolicName: com.liferay.osb.analytics.audit.header
-Bundle-Version: 1.0.0
